% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_design_classV2.5.R
\name{printDesign}
\alias{printDesign}
\alias{printDesign,sequential.design-method}
\title{Function that displays the features of the design}
\usage{
\S4method{printDesign}{sequential.design}(theObject)
}
\arguments{
\item{theObject}{An object of class \code{sequential.design}.}
}
\value{
Prints the details to the console.
}
\description{
Displays the look times, the base alpha and beta spending functions, and the noncentrality parameter
}
\examples{
design1<-sequential.design()
design1<-calcBoundaries(design1)
design1<-setAlphaspendfString(design1,"0.025*t^4")
design1<-setBetaspendfString(design1,"0.15*t^3")
printDesign(design1)
}
