% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_design_classV2.5.R
\name{summaryDesign}
\alias{summaryDesign}
\alias{summaryDesign,sequential.design-method}
\title{Function that shows the cumulative probabilities for efficacy and futility}
\usage{
\S4method{summaryDesign}{sequential.design}(theObject)
}
\arguments{
\item{theObject}{An object of class \code{sequential.design}.}
}
\value{
Prints a summary matrix to the console.
}
\description{
Shows the cumulative probability for efficacy and futility under the null and alternative hypotheses, the corresponding p-values, and the boundaries for Z at each look.
}
\examples{
design2 <- calcBoundaries(sequential.design())
# Summarize the design
summaryDesign(design2)
}
